#23 Restructure the repository
