+++ src/kmain.c
@@ -9,12 +9,11 @@
 #include <stddef.h>
 #include <stdint.h>
 
-#include "include/vga.h"
-#include "include/string.h"
-
+#include "libc/include/string.h"
+#include "libk/include/vga.h"
 
 void kernel_main(void) {
-  putstring(5, 2, "Works!\0");
   clear();
+  putstring(5, 2, "Works!\0");
   refresh();
 }
