+++ src/libc/include/string.h
