+++ src/libc/string.c
@@ -1,4 +1,4 @@
-#include "string.h"
+#include "include/string.h"
 
 void* memcpy(void *dest, const void *src, size_t count) {
   char *tmp = dest;
