+++ src/libk/font.c
@@ -1,4 +1,4 @@
-#include "font.h"
+#include "include/font.h"
 
 const uint16_t FONT_12x16[] = {
   /*
