+++ src/makefile
@@ -1,45 +1,98 @@
-AS = "$$HOME/tools/crc/bin/i686-elf-as"
-GCC = "$$HOME/tools/crc/bin/i686-elf-gcc"
+# REMARK mju 22-09-03:
+# Depending on where the toolchain was installed, the following paths to the compiler
+# or assembler need to be changed. To use Bash variables in the script, the dollar sign
+# $ introducing the variable must be masked as $$.
+
+# Assembler options
+AS = $$HOME/tools/crc/bin/i686-elf-as
+
+# C compiler options
+CC = $$HOME/tools/crc/bin/i686-elf-gcc
 CFLAGS = -std=gnu99 -ffreestanding -O2 -Wall -Wextra
+# Linker options
+LFLAGS = -ffreestanding -O2 -nostdlib
+LLIBS = -lgcc
 
+# Output directories
 BUILD_DIR = ../build
-OBJ_DIR = ../build/obj
-ISO_DIR = ../build/iso
-BOOT_DST = ../build/iso/boot
-GRUB_DIR = ../build/iso/boot/grub
+LIBC_OBJ_DIR = $(BUILD_DIR)/libc
+LIBK_OBJ_DIR = $(BUILD_DIR)/libk
+BOOT_OBJ_DIR = $(BUILD_DIR)/boot
+ISO_DIR = $(BUILD_DIR)/iso
 
-OBJECTS = $(OBJ_DIR)/boot.o $(OBJ_DIR)/kernel.o $(OBJ_DIR)/string.o $(OBJ_DIR)/font.o $(OBJ_DIR)/vga.o
+# Output filenames
+ISO = release.iso
+KERNEL = kernel.bin
 
-IMG = sys_img.bin
+# Object files
 
-.PHONY: mkdirs
-all: mkdirs $(OBJECTS) iso/*
-	$(GCC) -T linker.ld -o $(BUILD_DIR)/$(IMG) -ffreestanding -O2 -nostdlib $(OBJ_DIR)/* -lgcc
-	cp $(BUILD_DIR)/$(IMG) $(BOOT_DST)
-	cp iso/bootloader.cfg $(GRUB_DIR)/grub.cfg;
-	grub-mkrescue -o $(ISO_DIR)/system-img-$$(date +%F).iso $(ISO_DIR)
+LIBC_OBJ = \
+	$(LIBC_OBJ_DIR)/string.o
 
-mkdirs:
-	mkdir -p $(OBJ_DIR)
-	mkdir -p $(GRUB_DIR)
+LIBK_OBJ = \
+	$(LIBK_OBJ_DIR)/vga.o
+
+BOOT_OBJ = \
+	$(BOOT_OBJ_DIR)/boot.o
+
+KERNEL_IMG = $(BUILD_DIR)/kmain.o
+
+
+
+.PHONY: iso
+iso: kernel | $(ISO_DIR)
+	grub-mkrescue -o $(BUILD_DIR)/$(ISO) $(ISO_DIR)
+
+$(ISO_DIR):
+	mkdir -p $(ISO_DIR)/boot/grub
+	cp boot/grub.cfg $(ISO_DIR)/boot/grub/grub.cfg
+	cp $(BUILD_DIR)/$(KERNEL) $(ISODIR)/boot/$(KERNEL)
+
+.PHONY: kernel
+kernel: libc libk boot
+	$(CC) -c kmain.c -o $(KERNEL_IMG) $(CFLAGS)
+	$(CC) -T linker.ld -o $(BUILD_DIR)/$(KERNEL) $(LFLAGS) $(LLIBS)\
+		$(BOOT_OBJ) \
+		$(LIBC_OBJ) \
+		$(LIBK_OBJ) \
+		$(KERNEL_IMG)
 
 .PHONY: clean
 clean:
-	rm $(OBJ_DIR)/*
-	rm -r $(ISO_DIR)/*
-	rm $(BUILD_DIR)/$(IMG)
+	rm -r $(BUILD_DIR)/*
 
-$(OBJ_DIR)/boot.o: boot.s
-	$(AS) boot.s -o $(OBJ_DIR)/boot.o
+# Rules for libc
 
-$(OBJ_DIR)/kernel.o: kernel.c
-	$(GCC) -c kernel.c -o $(OBJ_DIR)/kernel.o $(CFLAGS)
+.PHONY: libc
+libc: $(LIBC_OBJ)
 
-$(OBJ_DIR)/string.o: include/string.c include/string.h
-	$(GCC) -c include/string.c -o $(OBJ_DIR)/string.o $(CFLAGS)
+$(LIBC_OBJ_DIR)/%.o: $(wildcard libc/*.c) | $(LIBC_OBJ_DIR)
+	$(CC) -c $< -o $@ $(CFLAGS)
 
-$(OBJ_DIR)/font.o: include/font.c include/font.h
-	$(GCC) -c include/font.c -o $(OBJ_DIR)/font.o $(CFLAGS)
+$(LIBC_OBJ_DIR):
+	mkdir -p $(LIBC_OBJ_DIR)
 
-$(OBJ_DIR)/vga.o: include/vga.c include/vga.h
-	$(GCC) -c include/vga.c -o $(OBJ_DIR)/vga.o $(CFLAGS)
+
+# Rules for libk
+
+.PHONY: libk
+libk: $(LIBK_OBJ)
+
+$(LIBK_OBJ_DIR)/%.o: $(wildcard libk/*.c) | $(LIBK_OBJ_DIR)
+	$(CC) -c $< -o $@ $(CFLAGS)
+
+$(LIBK_OBJ_DIR):
+	mkdir -p $(LIBK_OBJ_DIR)
+
+
+# Rules for boot
+
+.PHONY: boot
+boot: $(BOOT_OBJ_DIR)/boot.o
+
+$(BOOT_OBJ_DIR)/%.o: $(wildcard boot/*.s) | $(BOOT_OBJ_DIR)
+	$(AS) $< -o $@
+
+$(BOOT_OBJ_DIR):
+	mkdir -p $(BOOT_OBJ_DIR)
+
